[fix] サンプルとクラス定義にファイルを分割
