+++ IKS.gadget.js
@@ -1,4 +1,4 @@
-class IKSGadget {
+export class IKSGadget {
 
     constructor(){
         this.approval = false;
@@ -183,29 +183,6 @@ class IKSGadget {
         this.recoveryState = 0;
         return 0;
     }
-
-    // 指定座標の最寄り駅に連鎖的にチェックイン
-    async checkinAtNearest( origin, list ){
-
-        let point = origin;
-
-        const max = list.length;
-        for( let i=0; i<max; i++ ){
-            let res = searchNearestPoint( point, list );
-            list.splice(res.index, 1)
-
-            // checkin
-            let cInRes = null;
-            do{
-                this.form = (await this.getFormation()).contents;
-                cInRes = await this.checkin(coordTrick({lat:res.lat, lng:res.lng}), this.form[ getRandInt(this.form.length) ].name_en);
-                await this.delay(res.distance*2300);
-            }while(!cInRes.contents)
-
-            point.lng = res.lng;
-            point.lat = res.lat;
-        }
-    }
 }
 
 function getURIstr(params){
@@ -235,57 +212,3 @@ function mergeDeeply(target, source, opts) {
     }
     return result;
 }
-
-function searchNearestPoint( coord, list ){
-    const nearest = {
-        lng : 180,
-        lat : 180,
-        index : -1,
-        distance : 180
-    };
-
-    list.forEach((e, index)=>{
-        distance = Math.max( Math.sqrt(Math.abs(e.lng - coord.lng) * Math.abs(e.lat - coord.lat)), Math.abs(e.lng - coord.lng), Math.abs(e.lat - coord.lat));
-        if ( nearest.distance > distance ){
-            nearest.distance = distance;
-            nearest.lng = e.lng;
-            nearest.lat = e.lat;
-            nearest.index = index;
-        }
-    });
-
-    return nearest;
-}
-
-function getRandInt( max ){
-    return Math.floor(Math.random() * max);
-}
-
-function coordTrick( coord ){
-    const sign = [1, -1];
-    return {
-        lat : coord.lat + ( Math.random() / 1000 * sign[getRandInt(2)]),
-        lng : coord.lng + ( Math.random() / 1000 * sign[getRandInt(2)])
-    };
-}
-
-(async()=>{
-    const iks = new IKSGadget();
-    const list = await fetch("https://raw.githubusercontent.com/IKS-org/IKS.gadget/develop/coords.json?token=GHSAT0AAAAAABQK5Q3BTOMMBIKLOJNNDXW6YP5KVEA").then(r=>r.json());
-    const coords = list.map((e)=>{ return e.coord });
-
-    const start = {lng: 139.738477, lat: 35.752164};
-    const breakpoint = { lng: 138.853927, lat: 37.447787 }
-
-    let res = {};
-    while(true){
-        let res = searchNearestPoint( start, coords );
-        coords.splice(res.index, 1);
-        if(  res.lng == breakpoint.lng && res.lat == breakpoint.lat ) break;
-
-    }
-
-    await iks.agreement();
-    iks.checkinAtNearest( coordTrick(breakpoint), coords );
-
-})();
